chore(engine): reduce selected columns in acquisition query
Related to CAM-9838
