+++ engine/src/test/java/org/camunda/bpm/engine/test/jobexecutor/JobExecutorAcquireJobsByPriorityTest.java
@@ -24,6 +24,7 @@ import static org.junit.Assert.assertTrue;
 import java.util.List;
 
 import org.camunda.bpm.engine.impl.persistence.entity.AcquirableJobEntity;
+import org.camunda.bpm.engine.runtime.Job;
 import org.camunda.bpm.engine.test.Deployment;
 import org.junit.Before;
 import org.junit.Test;
@@ -66,19 +67,19 @@ public class JobExecutorAcquireJobsByPriorityTest extends AbstractJobExecutorAcq
     List<AcquirableJobEntity> acquirableJobs = findAcquirableJobs();
     assertEquals(20, acquirableJobs.size());
     for (int i = 0; i < 5; i++) {
-      assertEquals(10, acquirableJobs.get(i).getPriority());
+      assertEquals(10, findJobById(acquirableJobs.get(i).getId()).getPriority());
     }
 
     for (int i = 5; i < 10; i++) {
-      assertEquals(8, acquirableJobs.get(i).getPriority());
+      assertEquals(8, findJobById(acquirableJobs.get(i).getId()).getPriority());
     }
 
     for (int i = 10; i < 15; i++) {
-      assertEquals(5, acquirableJobs.get(i).getPriority());
+      assertEquals(5, findJobById(acquirableJobs.get(i).getId()).getPriority());
     }
 
     for (int i = 15; i < 20; i++) {
-      assertEquals(4, acquirableJobs.get(i).getPriority());
+      assertEquals(4, findJobById(acquirableJobs.get(i).getId()).getPriority());
     }
   }
 
@@ -92,6 +93,10 @@ public class JobExecutorAcquireJobsByPriorityTest extends AbstractJobExecutorAcq
     startProcess("jobPrioProcess", "task2", 5);
 
     // set some job priorities to NULL indicating that they were produced without priorities
+  }
+
+  protected Job findJobById(String id) {
+    return managementService.createJobQuery().jobId(id).singleResult();
   }
 
 }
